tflite: Support adding TMR data to only high-precision parameter buffers
We conducted bit-flip attack against various TFLite models and realized
that flipping single bits in the high-precision parameters can cause up
to 50% relative accuracy degradation. Other regions of the models are
not as vulnerable, so only protecting high-precision regions is
sufficient.

The redundant buffer arrays are organized to have the same indices as
the original data so it'd be easy for the inference code to load the
redundant data using the same index. Indicies that point to 8-bit
parameters have placeholder empty redundancy buffers.